MedAI-project/requirements.txt
torch>=2.2.0
transformers>=4.49.0
accelerate>=0.30.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
pyyaml>=6.0
scipy>=1.10.0
torchvision>=0.17.0
scikit-image>=0.22.0
qwen-vl-utils>=0.0.10